chore: move lint to pre-commit
